move pascal_voc --> voc
