+++ chainercv/datasets/__init__.py
@@ -1,8 +1,8 @@
 from chainercv.datasets.cub.cub_keypoint_dataset import CUBKeypointDataset  # NOQA
 from chainercv.datasets.cub.cub_label_dataset import CUBLabelDataset  # NOQA
 from chainercv.datasets.online_products.online_products_dataset import OnlineProductsDataset  # NOQA
-from chainercv.datasets.pascal_voc.voc_detection_dataset import VOCDetectionDataset  # NOQA
-from chainercv.datasets.pascal_voc.voc_semantic_segmentation_dataset import VOCSemanticSegmentationDataset  # NOQA
-from chainercv.datasets.pascal_voc.voc_utils import voc_detection_label_names  # NOQA
-from chainercv.datasets.pascal_voc.voc_utils import voc_semantic_segmentation_label_names  # NOQA
 from chainercv.datasets.transform_dataset import TransformDataset  # NOQA
+from chainercv.datasets.voc.voc_detection_dataset import VOCDetectionDataset  # NOQA
+from chainercv.datasets.voc.voc_semantic_segmentation_dataset import VOCSemanticSegmentationDataset  # NOQA
+from chainercv.datasets.voc.voc_utils import voc_detection_label_names  # NOQA
+from chainercv.datasets.voc.voc_utils import voc_semantic_segmentation_label_names  # NOQA
