+++ chainercv/datasets/voc/voc_detection_dataset.py
@@ -5,7 +5,7 @@ import xml.etree.ElementTree as ET
 
 import chainer
 
-from chainercv.datasets.pascal_voc import voc_utils
+from chainercv.datasets.voc import voc_utils
 from chainercv.utils import read_image
