+++ chainercv/datasets/voc/voc_semantic_segmentation_dataset.py
@@ -3,7 +3,7 @@ import os
 
 import chainer
 
-from chainercv.datasets.pascal_voc import voc_utils
+from chainercv.datasets.voc import voc_utils
 from chainercv.utils import read_image
