+++ chainercv/datasets/voc/voc_utils.py
